+++ apps/builder/app/shared/pages/index.ts
@@ -1,1 +1,2 @@
 export * from "./use-switch-page";
+export * from "./pages";
