+++ apps/builder/app/shared/pages/pages.test.ts
@@ -1,7 +1,8 @@
 import { describe, expect, test } from "@jest/globals";
 import type { Page, Pages } from "@webstudio-is/sdk";
-import { $pages, savePathInHistory } from "~/shared/nano-states";
+import { $pages } from "~/shared/nano-states";
 import { registerContainers } from "~/shared/sync";
+import { savePathInHistory } from "./pages";
 
 registerContainers();
